// client/src/models/bois/BoisDur.ts
export interface PostBoisDur {
  marque: string;
  fornisseur: string;
  n_fardou: string;
  pieces: number;
  long: number;
  larg: number;
  epaisseur: number;
  volume: number;
  prix_achat: number;
  prix_vente: number;
  date_creation: string;
}
export interface GetBoisDur {
  id: number;
  marque: string;
  fornisseur: string;
  n_fardou: string;
  pieces: number;
  long: number;
  larg: number;
  epaisseur: number;
  volume: number;
  prix_achat: number;
  prix_vente: number;
  date_creation: string;
}